+++ R/TaskSurv.R
@@ -15,65 +15,68 @@
 #' @export
 #' @examples
 #' library(mlr3)
-#' lung = mlr3misc::load_dataset("lung", package = "survival")
+#' lung = survival::lung
 #' lung$status = (lung$status == 2L)
-#' b = as_data_backend(lung)
 #' task = TaskSurv$new("lung",
-#'   backend = b, time = "time",
+#'   backend = lung, time = "time",
 #'   event = "status")
 #'
+#' # meta data
 #' task$target_names
 #' task$feature_names
 #' task$formula()
+#'
+#' # survival data
 #' task$truth()
+#' task$times()
+#' task$status()
+#' task$unique_times()
+#' task$unique_event_times()
+#' task$risk_set(time = 700)
+#'
 TaskSurv = R6::R6Class("TaskSurv",
   inherit = TaskSupervised,
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     #'
-    #' @param time (`character(1)`)\cr
-    #' Name of the column for event time if data is right censored, otherwise starting time if
-    #' interval censored.
-    #'
-    #' @param event (`character(1)`)\cr
-    #' Name of the column giving the event indicator.
-    #' If data is right censored then "0"/`FALSE` means alive (no event), "1"/`TRUE` means dead
-    #' (event). If `type` is `"interval"` then "0" means right censored, "1" means dead (event),
-    #' "2" means left censored, and "3" means interval censored. If `type` is `"interval2"` then
-    #' `event` is ignored.
-    #'
-    #' @param time2 (`character(1)`)\cr
-    #' Name of the column for ending time for interval censored data, otherwise ignored.
-    #'
-    #' @param type (`character(1)`)\cr
-    #' Name of the column giving the type of censoring. Default is 'right' censoring.
-    initialize = function(id, backend, time, event, time2,
-      type = c("right", "left", "counting", "interval", "interval2", "mstate")) {
+    #' @template param_time
+    #' @template param_event
+    #' @template param_time2
+    #' @template param_type
+    initialize = function(id, backend, time = "time", event = "event", time2,
+      type = c("right", "left", "interval", "counting", "interval2", "mstate")) {
+
       type = match.arg(type)
+
+      backend = as_data_backend(backend)
+
+      if (type != "interval2") {
+        c_ev = backend$.__enclos_env__$private$.data[, event, with = FALSE][[1]]
+        if (type == "mstate") {
+          assert_factor(c_ev)
+        } else if (type == "interval") {
+          assert_integerish(c_ev, lower = 0, upper = 3)
+        } else if (!is.logical(c_ev)) {
+          assert_integerish(c_ev, lower = 0, upper = 2)
+        }
+      }
+
+      private$.censtype = type
 
       if (type %in% c("right", "left", "mstate")) {
         super$initialize(
           id = id, task_type = "surv", backend = backend,
           target = c(time, event))
-      } else if (type %in% "interval2") {
-        super$initialize(
-          id = id, task_type = "surv", backend = backend,
-          target = c(time, time2))
-      } else {
+      } else if (type %in% c("interval", "counting")) {
         super$initialize(
           id = id, task_type = "surv", backend = backend,
           target = c(time, time2, event))
+      } else {
+        super$initialize(
+          id = id, task_type = "surv", backend = backend,
+          target = c(time, time2))
       }
-
-      if (type %nin% "interval2") {
-        event = self$data(cols = event)[[1L]]
-        if (!is.logical(event)) {
-          assert_integerish(event, lower = 0, upper = 3)
-        }
-      }
-
-      private$.censtype = type
     },
 
     #' @description
@@ -83,14 +86,16 @@ TaskSurv = R6::R6Class("TaskSurv",
     truth = function(rows = NULL) {
       # truth is defined as the survival outcome as a Survival object
       tn = self$target_names
+      ct = self$censtype
       d = self$data(rows, cols = self$target_names)
       args = list(time = d[[tn[1L]]], type = self$censtype)
-      if (self$censtype %in% "interval2") {
-        args$time2 = d[[tn[2L]]]
-      } else if (length(tn) == 2) {
+
+      if (ct %in% c("right", "left", "mstate")) {
         args$event = as.integer(d[[tn[2L]]])
-      } else {
+      } else if (ct %in% c("interval", "counting")) {
         args$event = as.integer(d[[tn[3L]]])
+        args$time2 = d[[tn[2L]]]
+      } else {
         args$time2 = d[[tn[2L]]]
       }
 
@@ -111,21 +116,87 @@ TaskSurv = R6::R6Class("TaskSurv",
     formula = function(rhs = NULL) {
       # formula appends the rhs argument to Surv(time, event)~
       tn = self$target_names
-      if (self$censtype %in% "interval2") {
-        lhs = sprintf("Surv(time = %s, time2 = %s, type = 'interval2')", tn[1L], tn[2L])
-      } else if (length(tn) == 2) {
+      if (length(tn) == 2) {
         lhs = sprintf("Surv(%s, %s, type = '%s')", tn[1L], tn[2L], self$censtype)
       } else {
         lhs = sprintf("Surv(%s, %s, %s, type = '%s')", tn[1L], tn[2L], tn[3L], self$censtype)
       }
       formulate(lhs, rhs %??% ".", env = getNamespace("survival"))
+    },
+
+    #' @description
+    #' Returns the (unsorted) outcome times.
+    #' @return `numeric()`
+    times = function(rows = NULL) {
+      truth = self$truth(rows)
+      if (self$censtype %in% c("interval", "counting", "interval2")) {
+        return(truth[, 1:2])
+      } else {
+        return(truth[, 1L])
+      }
+    },
+
+    #' @description
+    #' Returns the event indicator (aka censoring/survival indicator).
+    #' If `censtype` is `"right"` or `"left"` then `1` is event and `0` is censored.
+    #' If `censtype` is `"mstate"` then `0` is censored and all other values are different events.
+    #' If `censtype` is `"interval"` then `0` is right-censored, `1` is event, `2` is left-censored,
+    #' `3` is interval-censored.
+    #' See [survival::Surv].
+    #' @return `integer()`
+    status = function(rows = NULL) {
+      truth = self$truth(rows)
+      if (self$censtype %in% c("interval", "counting", "interval2")) {
+        status = truth[, 3L]
+      } else {
+        status = truth[, 2L]
+      }
+
+      as.integer(status)
+    },
+
+    #' @description
+    #' Returns the sorted unique outcome times for 'right', 'left', and 'mcstate'.
+    #' @return `numeric()`
+    unique_times = function(rows = NULL) {
+      if (self$censtype %in% c("interval", "counting", "interval2")) {
+        stop("Not implemented for 'interval', 'interval2', or 'counting', 'censtype'.")
+      }
+
+      sort(unique(self$times(rows)))
+    },
+
+    #' @description
+    #' Returns the sorted unique event (or failure) outcome times.
+    #' @return `numeric()`
+    unique_event_times = function(rows = NULL) {
+      if (self$censtype %in% c("interval", "counting", "interval2")) {
+        stop("Not implemented for 'interval', 'interval2', or 'counting', 'censtype'.")
+      }
+
+      sort(unique(self$times(rows)[self$status(rows) != 0]))
+    },
+
+    #' @description
+    #' Returns the `row_ids` of the observations 'at risk' (not dead or censored) at `time`.
+    #' @param time (`numeric(1)`) \cr Time to return risk set for, if `NULL` returns all `row_ids`.
+    #' @return `integer()`
+    risk_set = function(time = NULL) {
+      if (self$censtype %in% c("interval", "counting", "interval2")) {
+        stop("Not implemented for 'interval', 'interval2', or 'counting', 'censtype'.")
+      }
+
+      if (is.null(time)) {
+        self$row_ids
+      } else {
+        self$row_ids[self$times() >= time]
+      }
     }
   ),
 
   active = list(
     #' @field censtype `character(1)`\cr
-    #' Returns the type of censoring, one of "right", "left", "counting", "interval", "interval2"
-    #' or "mstate".
+    #' Returns the type of censoring, one of "right", "left", "counting", "interval", or "mstate".
     censtype = function() {
       return(private$.censtype)
     }