+++ man-roxygen/param_time.R
@@ -1,0 +1,3 @@
+#' @param time (`character(1)`)\cr
+#' Name of the column for event time if data is right censored, otherwise starting time if
+#' interval censored.