+++ man-roxygen/param_time2.R
@@ -1,0 +1,2 @@
+#' @param time2 (`character(1)`)\cr
+#' Name of the column for ending time for interval censored data, otherwise ignored.