+++ tic.R
@@ -4,7 +4,7 @@ do_package_checks()
 if (ci_on_ghactions() && ci_has_env("BUILD_PKGDOWN")) {
   
   get_stage("before_deploy") %>%
-    add_step(step_setup_push_deploy(branch = "master", orphan = FALSE))
+    add_step(step_setup_push_deploy(branch = "main", orphan = FALSE))
   
   # creates pkgdown site and pushes to gh-pages branch
   # only for the runner with the "BUILD_PKGDOWN" env var set
@@ -17,6 +17,3 @@ if (ci_on_ghactions() && ci_has_env("BUILD_PKGDOWN")) {
     add_code_step(rmarkdown::render("README.Rmd")) %>%
     add_step(step_do_push_deploy(commit_paths = c("README.md")))
 }
-
-get_stage("after_success") %>%
-  add_code_step(system("curl -s https://raw.githubusercontent.com/mlr-org/mlr3orga/master/trigger-mlr3book.sh | bash"))